+++ src/app.ts
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2026 Marcel Petrick <mail@marcelpetrick.it>
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-import { Engine } from '@babylonjs/core';
+import { Engine } from '@babylonjs/core/Engines/engine';
 import pkg from '../package.json';
 import { Audio, type FlightSound } from './audio';
 import { FLYER_SPEED } from './core/flyer';