+++ src/render/themes.ts
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2026 Marcel Petrick <mail@marcelpetrick.it>
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-import { Color3 } from '@babylonjs/core';
+import { Color3 } from '@babylonjs/core/Maths/math.color';
 
 export interface Theme {
   id: string;